requirements.txt
supabase==2.9.1
pyyaml
Pillow
requests